[#72] chore: widget, widgetview 폴더링 분배
